% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_cosine_similarities}
\alias{compute_cosine_similarities}
\title{Compute cosine similarities}
\usage{
compute_cosine_similarities(matrices_list, threshold)
}
\arguments{
\item{matrices_list}{A list of matrices where the row names are document IDs and columns are embeddings.}

\item{threshold}{A threshold value for the cosine similarity.}
}
\value{
A list of data frames, each containing the document ID pairs and their cosine similarity values.
}
\description{
This function computes the cosine similarity between the first row and all other rows of each matrix in a list.
}
