\name{coorsim-package}
\alias{coorsim-package}
\alias{coorsim}
\docType{package}
\title{
\packageTitle{coorsim}
}
\description{
\packageDescription{coorsim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{coorsim}
\packageIndices{coorsim}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{coorsim}

Maintainer: \packageMaintainer{coorsim}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
