% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_detect_communities.R
\name{coorsim_detect_groups}
\alias{coorsim_detect_groups}
\title{Detect Communities of Coordinated Accounts}
\usage{
coorsim_detect_groups(
  simdt,
  user_data,
  account_id = "account_id",
  account_name = "account_name",
  other_user_vars = NULL,
  edge_weight = NULL,
  cluster_method = "FSA_V",
  resolution = 1,
  theta = 0.3,
  return_post_dt = TRUE,
  verbose = TRUE,
  seed = 42
)
}
\arguments{
\item{simdt}{A `data.table` containing the network edges, with columns for account IDs 
(`account_id`, `account_id_y`) and optional columns for `post_id` and `time`.}

\item{user_data}{A `data.table` of account-level data with at least one column for `account_id`. 
Additional columns can be specified to be included in the output.}

\item{account_id}{Name of the column in `user_data` that identifies accounts. Defaults to `"account_id"`.}

\item{account_name}{Name of the column in `user_data` for account names or labels. Defaults to `"account_name"`.}

\item{other_user_vars}{Optional character vector of additional column names in `user_data` 
to include in the output.}

\item{edge_weight}{Optional numeric threshold for filtering the edge list. 
Only edges with a weight greater than or equal to this value will be included.}

\item{cluster_method}{Character, the algorithm to use for community detection. 
Options are `"louvain"` or `"FSA_V"`. Defaults to `"FSA_V"`.}

\item{resolution}{Numeric resolution parameter for the Louvain algorithm. 
Controls the size of detected communities. Higher values create smaller communities. Defaults to `1`.}

\item{theta}{Numeric, specifies the threshold for the FSA_V algorithm to control 
the inclusion of edges. Higher values prioritize stronger edges in community construction. Defaults to `0.7`.}

\item{return_post_dt}{Logical, whether to return a `data.table` with post data merged with community labels. 
Defaults to `TRUE`.}

\item{verbose}{Logical, whether to display progress messages. Defaults to `TRUE`.}

\item{seed}{Numeric, the random seed to ensure reproducibility. Defaults to `42`.}
}
\value{
A named list containing:
  \describe{
    \item{\code{graph}}{An `igraph` object representing the network graph created from the filtered edge list.}
    \item{\code{communities}}{An `igraph::communities` object representing the detected communities.}
    \item{\code{edge_list}}{A filtered `data.table` of edges used for community detection.}
    \item{\code{node_list}}{A `data.table` with node membership information, including account ID, community assignment, and algorithm used.}
    \item{\code{post_data}}{A `data.table` containing post-level data with community information, only returned if \code{return_post_dt} is `TRUE`.}
  }
}
\description{
This function identifies communities of coordinated accounts in a co-similarity network. 
It supports two community detection algorithms-the classic "Louvain" algorithme and the Focal Structures Algorithm "FSA_V"
proposed by Weber & Neumann (2021), based on Sen et al. (2016). It can filter communities based on a minimum edge weight threshold.
}
