% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_coorsimilarity.R
\name{coorsim_prepare_data}
\alias{coorsim_prepare_data}
\title{Prepare Data for Correlation Similarity Analysis}
\usage{
coorsim_prepare_data(
  data,
  vector_matrix = NULL,
  time_window = 60,
  post_id = NULL,
  account_id = NULL,
  time = NULL,
  content = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame or data table containing the posts data.}

\item{vector_matrix}{A matrix with vector represenations for social media items. The row names should correspond to post IDs.}

\item{time_window}{An integer specifying the time window (in seconds) within which to detect similar posts. Default is 60.}

\item{post_id}{A character string representing the alternative name for the post ID column in the data (if not named "post_id").}

\item{account_id}{A character string representing the alternative name for the account ID column in the data (if not named "account_id").}

\item{time}{A character string representing the alternative name for the time column in the data (if not named "time").}

\item{content}{A character string representing the alternative name for the content column in the data (if not named "content").}

\item{verbose}{A logical value indicating whether to display progress messages. Default is TRUE.}
}
\value{
A data table with the prepared data, including necessary transformations and subsetting.
}
\description{
This function prepares data for correlation similarity analysis by performing various checks and transformations 
on the input data and vector matrix. It ensures that the necessary columns are present, converts data types, and 
subsets the data as required.
}
