% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_ntile}
\alias{filter_ntile}
\title{Filter Data by Quantile Threshold}
\usage{
filter_ntile(df, var_field, probs)
}
\arguments{
\item{df}{A data frame or data.table to filter.}

\item{var_field}{Character or integer indicating the column to filter by. If character, it should match the column name; 
if integer, it should match the column index.}

\item{probs}{Numeric value between 0 and 1 indicating the quantile threshold for filtering. Rows with values 
in `var_field` greater than or equal to this quantile are retained.}
}
\value{
The filtered data frame or data.table containing rows where the specified variable is at or above the specified quantile.
}
\description{
This function filters a data frame or data.table by retaining rows where the specified variable is 
greater than or equal to a given quantile threshold. It supports filtering by the variable's name 
or by its index position.
}
