% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_get_embeddings.R
\name{get_embeddings}
\alias{get_embeddings}
\title{Retrieve Text Embeddings Using Hugging Face Model via Python Script}
\usage{
get_embeddings(
  data,
  post_id = NULL,
  time = NULL,
  content = NULL,
  model_name = "sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2",
  batch_size = 16L,
  max_length = 512L,
  use_fp16 = TRUE,
  trust_remote_code = TRUE,
  verbose = TRUE,
  uv_cache_dir = NULL,
  models_dir = NULL
)
}
\arguments{
\item{data}{A `data.frame` or `data.table` containing the input text data.}

\item{post_id}{Character. Column name in `data` containing unique post identifiers. If `NULL`, 
the function attempts to detect it automatically.}

\item{time}{Character. Column name in `data` containing timestamps. If `NULL`, the function 
attempts to detect it automatically.}

\item{content}{Character. Column name in `data` containing text content to embed. If `NULL`, 
the function attempts to detect it automatically.}

\item{model_name}{The Hugging Face model to use for embeddings (default is the CPU-friendly, multilingual model "sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2").
Other useful, more complex models to consider are, e.g.:
\itemize{
  \item Social-media specific, multilingual: "Twitter/twhin-bert-base"
  \item Multilingual, lightweight model, trained on wikipedia: "distilbert/distilbert-base-multilingual-cased"
  }}

\item{batch_size}{Integer. Batch size for processing texts (default is 32L).}

\item{max_length}{Integer. Maximum sequence length for tokenization (default is 512L).}

\item{use_fp16}{Logical, whether to use fp16 precision on GPU if available (default is FALSE).}

\item{trust_remote_code}{Logical, whether to trust remote code from huggingface or not (default is TRUE). Necessary for Alibaba gte.}

\item{verbose}{Logical. If `TRUE`, prints progress messages.}

\item{uv_cache_dir}{Character (optional). Directory used by uv to install python libraries, passed on to initialize_coorsim.}

\item{models_dir}{Character (optional).  Directory used to cache huggingface models, passed on to initialize_coorsim.}
}
\value{
A matrix with rownames corresponding to the 'id' column and computed embeddings as columns.
}
\description{
Retrieve Text Embeddings Using Hugging Face Model via Python Script
}
