% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_install_conda_coorsim.R
\name{initialize_conda_coorsim}
\alias{initialize_conda_coorsim}
\title{Initialize the Coorsim Conda Environment}
\usage{
initialize_conda_coorsim(
  python_version = "3.13",
  conda_path = NULL,
  conda_env_path = NULL,
  conda_env_name = "conda-coorsim",
  ask = TRUE,
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{python_version}{Character. The Python version to install in the Conda environment.
Defaults to `"3.13"`.}

\item{conda_path}{Character. The path where Miniconda should be installed.
If `NULL`, Miniconda is installed in the default location.}

\item{conda_env_path}{Character. The path where the Conda environment should be created.
If `NULL`, the environment is created in the default Conda environments directory.}

\item{conda_env_name}{Character. The name of the Conda environment to create or use.
Defaults to `"conda-coorsim"`.}

\item{ask}{Logical. If `TRUE`, prompts the user for confirmation before installing Miniconda.
Defaults to `TRUE`.}

\item{force}{Logical. If `TRUE`, forces reinstallation of dependencies even if they are already installed.
Defaults to `FALSE`.}

\item{verbose}{Logical. If `TRUE`, prints progress messages.}
}
\value{
Invisibly returns `NULL`. The function is executed for its side effects
of verifying and setting up the Conda environment and dependencies.
}
\description{
This function ensures that the required Conda environment for Coorsim is set up,
activated, and properly configured. If the environment is not found, it installs Conda,
sets up the environment, and verifies the installation of required dependencies.
}
\details{
This function:
\itemize{
  \item Checks if the Coorsim Conda environment is already initialized.
  \item Installs and activates Conda if necessary.
  \item Detects and verifies the active Conda environment.
  \item Confirms the installation of PyTorch and Transformers.
  \item Checks CUDA availability and version.
  \item Sources the required Python script from the package.
  \item Sets a global option indicating that the Coorsim Conda environment has been initialized.
}

If the Conda environment is already initialized, the function does nothing.
Otherwise, it installs and sets up all required dependencies before making them available.
}
