% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_py_transformers.R
\name{install_transformers}
\alias{install_transformers}
\title{Install Python, PyTorch, and Transformers}
\usage{
install_transformers(
  py_version = "3.11:latest",
  ask = interactive(),
  force = FALSE,
  pytorch_cuda_url = "https://download.pytorch.org/whl/cu124",
  ...
)
}
\arguments{
\item{py_version}{Character string. The version of Python to install, specified in 
the format \code{"<major>.<minor>:<latest>"}. Defaults to \code{"3.11:latest"}.}

\item{ask}{Logical. Whether to prompt the user before installing Python if not found.
Defaults to \code{interactive()}.}

\item{force}{Logical. If \code{TRUE}, forces the reinstallation of PyTorch and Transformers
even if they are already installed. Defaults to \code{FALSE}.}

\item{pytorch_cuda_url}{Character string. The URL for the PyTorch CUDA wheel index.
Defaults to \code{"https://download.pytorch.org/whl/cu124"}.}

\item{...}{Additional arguments passed to \code{reticulate::install_python()}.}
}
\description{
Installs a Python virtual environment, PyTorch, and the Transformers library. This function
handles Python installation if not already available and ensures that the necessary Python
packages are installed in the virtual environment.
}
\details{
The function checks for the existence of a virtual environment specified by the 
\code{COORSIM_PYTHON} environment variable. If not found, it creates the environment and installs
Python. Subsequently, it installs PyTorch and Transformers, using the appropriate version
based on CUDA availability. This ensures a seamless setup for using these libraries in R.
}
