% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_get_embeddings.R
\name{load_h5_embeddings}
\alias{load_h5_embeddings}
\title{Load Embeddings from an HDF5 (.h5) File}
\usage{
load_h5_embeddings(path, ids_subset = NULL, verbose = TRUE)
}
\arguments{
\item{path}{Character. Path to the `.h5` file containing embeddings.}

\item{ids_subset}{Character. Vector of IDs to subset the embedding matrix. Default is `NULL`.}

\item{verbose}{Logical. If `TRUE`, displays progress messages. Default is `TRUE`.}
}
\value{
A numeric matrix where rows correspond to post IDs and columns represent embedding dimensions. If no matching post IDs are found, returns `NULL`.
}
\description{
This function loads precomputed embeddings stored in an `.h5` file and retrieves only the relevant embeddings matching post IDs present in a dataset.
}
\details{
The function checks whether the `.h5` file contains a `metadata/post_id` dataset and loads the corresponding embeddings for matching post IDs.
}
