% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_predict_community_relevance.R
\name{load_xgb_model}
\alias{load_xgb_model}
\title{Load the default XGBoost model included in the package}
\usage{
load_xgb_model()
}
\value{
An `xgb.Booster` object.
}
\description{
This function loads a pre-trained `xgboost::xgb.Booster` model stored in `inst/extdata/final_model.xgb`.
}
