% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_coordinated_posts}
\alias{plot_coordinated_posts}
\title{Plot Coordinated Posts by Community and Time}
\usage{
plot_coordinated_posts(
  network_data,
  unit = "hour",
  n_communities = NULL,
  quantile = NULL,
  param_info = NULL,
  verbose = TRUE,
  legend = TRUE,
  title_prefix = NULL
)
}
\arguments{
\item{network_data}{A list containing data.tables with network data, where `post_data` must be one of the elements.}

\item{unit}{Character string specifying the time unit for flooring dates, defaulting to `"hour"`. Accepts units recognized by `lubridate::floor_date`.}

\item{n_communities}{Integer specifying the number of top communities to plot. Defaults to `NULL`, in which case all communities are included.}

\item{quantile}{Numeric value between 0 and 1 specifying the quantile threshold for filtering communities by size. 
Communities below this quantile threshold are excluded.}

\item{param_info}{Optional parameter info string that appears in the plot caption, providing context for the plot, such as parameter settings.}

\item{verbose}{Logical indicating whether to print status updates during execution. Defaults to `TRUE`.}

\item{legend}{Logical value indicating whether to display a legend for the plot. Defaults to `TRUE`.}

\item{title_prefix}{Optional character string used as a prefix for the plot title.}
}
\value{
A `ggplot` object visualizing the number of coordinated posts per community over time. 
If `n_communities` is specified, the plot is limited to the top `n_communities` by size.
}
\description{
This function generates a plot of coordinated posts over time by community, with options to filter by the top `n` communities
based on community size or post frequency. It utilizes a stacked bar chart to show the volume of posts per community, 
aggregated by the specified time unit.
}
