% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_predict_community_relevance.R
\name{predict_community_relevance}
\alias{predict_community_relevance}
\title{Predict ranking scores for communities using a trained XGBoost model}
\usage{
predict_community_relevance(groups_data, model = NULL, verbose = TRUE)
}
\arguments{
\item{groups_data}{A named list returned by `coorsim::get_community_metrics()`, including elements `community_metrics` and `params`.}

\item{model}{Trained `xgboost::xgb.Booster` object (optional). If `NULL`, loads default model from package.}

\item{verbose}{Logical. If `TRUE`, prints messages during preprocessing. Default is `TRUE`.}
}
\value{
A `groups_data` list, with `pred_relevance` -- a indicator ranging from 0 = irrelevant, to 6 = most relevatnt, `pred_relevance_score`, 
`pred_relevance_rank` added to `community_metrics` and `post_data`.
}
\description{
Predict ranking scores for communities using a trained XGBoost model
}
