% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_community_labels.R
\name{prepare_account_texts}
\alias{prepare_account_texts}
\title{Prepare and Clean Account Texts}
\usage{
prepare_account_texts(
  groups_data,
  sample_n = 5,
  sep = " + ",
  trunc_width = 100,
  min_n_char = 5,
  verbose = TRUE,
  seed = 42
)
}
\arguments{
\item{groups_data}{A list containing `post_data` and `node_list` data tables, where `post_data` includes posts associated with each account, and `node_list` includes account metadata.}

\item{sample_n}{Numeric. The number of posts to sample per account. Defaults to 5.}

\item{sep}{Character. A separator to use between sampled posts in the aggregated output. Defaults to `" + "`.}

\item{trunc_width}{Numeric. The maximum character width for truncating texts. Set to `NULL` for no truncation. Defaults to 100.}

\item{min_n_char}{Numeric. The minimum number of characters a post must have to be included in sampling. Defaults to 5.}

\item{verbose}{Logical. If `TRUE`, provides informational messages about each processing step. Defaults to `TRUE`.}

\item{seed}{Numeric. A seed for random sampling, ensuring reproducibility. Defaults to 42.}
}
\value{
A modified version of `groups_data`, with a new `account_text_prep` field in the `node_list`, which consolidates the cleaned name, location, bio, and sampled posts for each account.
}
\description{
This function processes account-related text data in a given `groups_data` object. It cleans post content, 
samples a specified number of posts per account, and prepares a consolidated text output containing account 
names, bios, and locations. This output is added to the `node_list` in `groups_data`.
}
\details{
The function proceeds through the following steps:
1. Cleans the post content by removing URLs, mentions, and special characters, and truncates texts if `trunc_width` is specified.
2. Filters posts by `min_n_char` and samples up to `sample_n` posts per account.
3. Merges the sampled post content with `node_list` in `groups_data`.
4. Cleans and truncates account metadata, such as names, descriptions, and locations.
5. Creates a consolidated text output for each account in the `account_text_prep` field.
}
