% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{query_embedding}
\alias{query_embedding}
\title{Query Embedding Subset}
\usage{
query_embedding(m, post_id_lists)
}
\arguments{
\item{m}{A NumericMatrix containing the embedding data with row names.}

\item{post_id_lists}{A List of CharacterVector objects, where each CharacterVector
represents a set of post IDs.}
}
\value{
A List of NumericMatrix objects. Each matrix in the list contains the
rows from the input matrix \code{m} that correspond to the post IDs in the
corresponding element of \code{post_id_lists}.
}
\description{
This function takes a NumericMatrix and a list of post ID lists, 
and returns a list of NumericMatrix subsets where each subset corresponds to 
the rows in the input matrix matched by the post IDs in each list.
}
