% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_num_threads}
\alias{set_num_threads}
\title{Set number of threads}
\usage{
set_num_threads(threads)
}
\arguments{
\item{threads}{Integer, setting the number of threads for parallel processing.}
}
\description{
Set number of threads
}
