% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_merging_filtering_functions.R
\name{filter_groups_data}
\alias{filter_groups_data}
\title{Filter and reduce a `groups_data` object by edge, time, similarity, or node-based constraints}
\usage{
filter_groups_data(
  groups_data,
  edge_weight = NULL,
  time_window = NULL,
  min_simil = NULL,
  min_comm_size = NULL,
  min_degree = NULL,
  communities_index = NULL,
  quantile_accountwise = NULL,
  quantile_postwise = NULL,
  top_n_accountwise = NULL,
  top_n_postwise = NULL,
  stringsimil_cutoff = NULL,
  stringsimil_method = "cosine",
  stringsimil_ngram = 3,
  nthread = 10,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{groups_data}{A named list from `coorsim::coorsim_detect_groups()` with `graph`, `communities`, `edge_list`, `node_list`, etc.}

\item{edge_weight}{Numeric. Minimum edge weight to retain.}

\item{time_window}{Integer. Max seconds between coordinated posts (triggers re-detection).}

\item{min_simil}{Numeric. Minimum cosine similarity between posts (triggers re-detection).}

\item{min_comm_size}{Integer. Minimum size of communities to keep.}

\item{min_degree}{Integer. Minimum node degree to retain.}

\item{communities_index}{Integer vector. Community indices to keep.}

\item{quantile_accountwise}{Numeric. Top quantile of communities by account count.}

\item{quantile_postwise}{Numeric. Top quantile of communities by post count.}

\item{top_n_accountwise}{Integer. Top N communities by account count.}

\item{top_n_postwise}{Integer. Top N communities by post count.}

\item{stringsimil_cutoff}{Numeric or `NULL`. If set, activates an additional filter by ngram-string-similarity. 
Pairs where the geometric mean of `similarity * string_similarity` are below the cutoff are removed.}

\item{stringsimil_method}{Character. String similarity method used by `stringdist::stringsim()`. Default is `"cosine"`.}

\item{stringsimil_ngram}{Integer. Size of n-grams used for calculating string similarity. Defaults to `3`.}

\item{nthread}{Integer. Number of threads used by `stringdist::stringsim()`. Defaults to `10`.}

\item{verbose}{Logical. Whether to print progress updates using `cli::cli_inform()`.}

\item{...}{Additional arguments passed to `coorsim_detect_groups()`, triggered after further filtering of `sim_dt`.}
}
\value{
A filtered `groups_data` list with a new `filter` element listing all applied parameters.
}
\description{
This function filters a `groups_data` object (typically returned by `coorsim::coorsim_detect_groups()`)
according to user-specified thresholds. If stricter time or similarity constraints are provided,
it will re-run group detection. At the end, it attaches a `filter` list to the returned object
that records all active filter parameters.
}
