% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_get_embeddings.R
\name{get_embeddings}
\alias{get_embeddings}
\title{Retrieve Text Embeddings Using Hugging Face Model via Python Script}
\usage{
get_embeddings(
  data,
  post_id = NULL,
  time = NULL,
  content = NULL,
  model_name = "sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2",
  batch_size = 16L,
  max_length = 512L,
  use_fp16 = TRUE,
  python_version = "3.13",
  conda_path = NULL,
  conda_env_path = NULL,
  conda_env_name = "conda-coorsim",
  ask = TRUE,
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A `data.frame` or `data.table` containing the input text data.}

\item{post_id}{Character. Column name in `data` containing unique post identifiers. If `NULL`, 
the function attempts to detect it automatically.}

\item{time}{Character. Column name in `data` containing timestamps. If `NULL`, the function 
attempts to detect it automatically.}

\item{content}{Character. Column name in `data` containing text content to embed. If `NULL`, 
the function attempts to detect it automatically.}

\item{model_name}{The Hugging Face model to use for embeddings (default is the CPU-friendly, multilingual model "sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2").
Other useful, more complex models to consider are, e.g.:
\itemize{
  \item Social-media specific, multilingual: "Twitter/twhin-bert-base"
  \item Multilingual, lightweight model, trained on wikipedia: "distilbert/distilbert-base-multilingual-cased"
  }}

\item{batch_size}{Integer. Batch size for processing texts (default is 32L).}

\item{max_length}{Integer. Maximum sequence length for tokenization (default is 512L).}

\item{use_fp16}{Logical, whether to use fp16 precision on GPU if available (default is FALSE).}

\item{python_version}{Character. The Python version to install in the Conda environment.
Defaults to `"3.13"`.}

\item{conda_path}{Character. The path where Miniconda should be installed.
If `NULL`, Miniconda is installed in the default location.}

\item{conda_env_path}{Character. The path where the Conda environment should be created.
If `NULL`, the environment is created in the default Conda environments directory.}

\item{conda_env_name}{Character. The name of the Conda environment to create or use.
Defaults to `"conda-coorsim"`.}

\item{ask}{Logical. If `TRUE`, prompts the user for confirmation before installing Miniconda.
Defaults to `TRUE`.}

\item{force}{Logical. If `TRUE`, forces reinstallation of dependencies even if they are already installed.
Defaults to `TRUE`.}

\item{verbose}{Logical. If `TRUE`, prints progress messages.}
}
\value{
A matrix with rownames corresponding to the 'id' column and computed embeddings as columns.
}
\description{
Retrieve Text Embeddings Using Hugging Face Model via Python Script
}
