% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_py_transformers.R
\name{install_torch}
\alias{install_torch}
\title{Install PyTorch with GPU or CPU Support}
\usage{
install_torch(pytorch_cuda_url = "https://download.pytorch.org/whl/cu124")
}
\arguments{
\item{pytorch_cuda_url}{Character string. The URL for the PyTorch CUDA wheel index. 
Defaults to \code{"https://download.pytorch.org/whl/cu124"}.}
}
\description{
Automatically installs PyTorch, TorchVision, and TorchAudio with the appropriate configuration
for GPU (CUDA) or CPU-only support, based on the system's capabilities.
}
\details{
The function checks for CUDA availability on the system and installs the corresponding
PyTorch version. If CUDA is not available, the CPU-only version is installed. The function
uses \code{reticulate} to manage Python virtual environments and ensure proper installation.
}
