% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_communities}
\alias{plot_communities}
\title{Plot a Community Graph with Enhanced Visualization}
\usage{
plot_communities(
  network_data,
  edge_weight_threshold = 5,
  community_size_threshold = 5,
  palette_option = "A",
  end_color = 0.9
)
}
\arguments{
\item{network_data}{A list containing an igraph graph, and a node_list as resulting from coorsim::detect_communities.}

\item{edge_weight_threshold}{Numeric, the minimum weight for edges to be included in the plot.}

\item{community_size_threshold}{Numeric, the minimum size of components to be included.}

\item{palette_option}{Character, the viridis palette option for coloring communities (default is "A").}

\item{end_color}{Numeric, the end value for color intensity in the viridis palette (default is 0.9).}
}
\value{
A ggplot2 object representing the community graph.
}
\description{
This function takes a network object with a graph and node list and 
creates a community graph visualization. Nodes with degree <= 1 
are removed, and the graph is further filtered based on edge weights 
and component size. Community colors are assigned using the viridis palette.
}
