% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_communities}
\alias{plot_communities}
\title{Plot communities in a network graph}
\usage{
plot_communities(
  network_data,
  edge_weight_threshold = NULL,
  component_size_threshold = NULL,
  palette_option = "A",
  end_color = 1,
  start_color = 0,
  title_prefix = NULL,
  n_top_nodes = 10
)
}
\arguments{
\item{network_data}{A list containing the graph object (`$graph`) and node list (`$node_list`). Optionally, it can include 
labelled communities (`$labelled_communities`), parameter list (`$param_list`), and filter information (`$filter_info`).}

\item{edge_weight_threshold}{Numeric. If provided, edges with a weight below this threshold will be removed. Defaults to `NULL`.}

\item{component_size_threshold}{Numeric. If provided, components smaller than this threshold will be removed. Defaults to `NULL`.}

\item{palette_option}{Character. Specifies the Viridis palette option to use. Defaults to "A".}

\item{end_color}{Numeric. The end point for the color gradient (between 0 and 1). Defaults to 1.}

\item{start_color}{Numeric. The start point for the color gradient (between 0 and 1). Defaults to 0.}

\item{title_prefix}{Character. A prefix to be added to the title of the plot. Defaults to `NULL`.}

\item{n_top_nodes}{Numeric. The number of top nodes by degree to display labels for within each community. Defaults to 10.}
}
\value{
A `ggplot` object representing the plotted network graph with communities, node labels, and boundary highlights.
}
\description{
This function generates a plot of communities in a network graph using a stress layout. The graph
is filtered based on component size and edge weight thresholds. Node colors and soft community boundaries 
are assigned based on community membership, and node labels for the top nodes by degree are displayed.
}
